+++ src/rclone/rclone_cli.cpp
@@ -250,42 +250,6 @@ void RcloneCli::lsjson(const std::string& remote_path,
     });
 }
 
-void RcloneCli::copy_files(const std::string& src_dir,
-                            const std::string& dst_dir,
-                            const std::vector<std::string>& includes,
-                            AsyncCallback<std::monostate> callback) {
-    std::vector<std::string> args = {"copy", src_dir, dst_dir};
-    for (auto& inc : includes) {
-        args.push_back("--include");
-        args.push_back(inc);
-    }
-    run_command(std::move(args), [callback = std::move(callback)](
-        const std::string&, const std::string& err, int code) {
-        if (code != 0)
-            callback(std::unexpected("copy failed: " + err));
-        else
-            callback(std::monostate{});
-    });
-}
-
-void RcloneCli::move_files(const std::string& src_dir,
-                            const std::string& dst_dir,
-                            const std::vector<std::string>& includes,
-                            AsyncCallback<std::monostate> callback) {
-    std::vector<std::string> args = {"move", src_dir, dst_dir};
-    for (auto& inc : includes) {
-        args.push_back("--include");
-        args.push_back(inc);
-    }
-    run_command(std::move(args), [callback = std::move(callback)](
-        const std::string&, const std::string& err, int code) {
-        if (code != 0)
-            callback(std::unexpected("move failed: " + err));
-        else
-            callback(std::monostate{});
-    });
-}
-
 void RcloneCli::delete_files(const std::string& dir,
                                const std::vector<std::string>& includes,
                                AsyncCallback<std::monostate> callback) {